[instSegm]: areaTH use unified scale.

+++ inc/vsn/vsnLib.h
@@ -475,7 +475,7 @@ namespace vsn{
                 HSV c0,c1;
             }; Filter filter;
             float blurSz = 3;
-            float areaTH = 20*10;
+            double areaTH = 20*10;
             bool en_imo = false;
             bool enShow = false;
             // in case color filter around red,
